feat(scripts): add `request.body.getComputed()`
Returns the `string` request body as sent via curl; with variables substituted,
or `undefined` if there is no body.

Very useful if you want to see the request body as it was sent to the server.

The `tryGetSubstituted` method will substitute variables with their values,
but leave the rest of the body as is.

If you have a GraphQL query in the body, for example, the `getComputed`
method will show the query as it was sent to the server,
which is quite different from the substituted version.

As an example, if you have a request body like this:

```graphql
query getRestClient($name: String!) {
  restclient(name: $name) {
    id
    name
    editorsSupported {
      name
    }
  }
}

{
  "variables": {
    "name": "{{ENV_VAR_CLIENT_NAME}}"
  }
}
```

Then the `getComputed` method will
return the body as it was sent to the server:

```json
{"query": "query getRestClient($name: String!) { restclient(name: $name) { id name editorsSupported { name } } } ", "variables": {"variables": {"name": "kulala"}}}
```

whereas the `tryGetSubstituted` method will
return the body with variables substituted as seen in your script:

```graphql
query getRestClient($name: String!) {
  restclient(name: $name) {
    id
    name
    editorsSupported {
      name
    }
  }
}

{
  "variables": {
    "name": "kulala"
  }
}
```

The `getComputed` method is always `undefined` for binary bodies.